+++ HISTORY.rst
@@ -2,11 +2,13 @@
 History
 =======
 
+
 In progress
 -----------
 
 * Allow networkx 2+ to be used
 * Add NameResolver application
+* Add enum for connector relation types
 
 
 0.2.0 (2018-05-30)